requirements.txt
cx_Freeze==7.1.1
cx_Logging==3.2.0
intelhex==2.3.0
lief==0.14.1
pyserial==3.5
